services/arso-service/requirements.txt
psycopg2-binary
fastapi
uvicorn
pika
python-dateutil
uuid-utils